Use GCC dependency generation in Makefiles
This causes source files which depend on headers to be rebuilt
when the header changes. `-MMD` generates the dependencies in
`*.d` files, excluding system headers. `-MP` adds fake rules
for headers to the dependencies file so build doesn't break if
a header is removed. Update `.gitignore` to ignore new `*.d`
files.